chore(atomix): refactor raft commit listener and remove state machine * commit listener only notifies the index instead of the entry. So it does not have to re-read the entries to notify listeners. * Rename ZeebeRaftStateMachine to LogCompactor
